func_stage/requirements.txt
azure-functions
azure-eventhub
azure-identity
azure-keyvault-secrets
opencensus-ext-azure
# shared-module dependencies required by DriftMonitor
requests>=2.31.0
python-dotenv>=1.0.0
pyodbc>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
azure-ai-ml>=1.12.0
azure-core>=1.30.0
matplotlib>=3.7.0
